requirements.txt
fastapi
uvicorn
streamlit
requests
pydantic
requests
python-dotenv
